refactor(stations): reworked Database page to Stations page, moved button groups to Settings page
