+++ src/renderer/src/features/StationsPage/StationsPage.tsx
@@ -1,0 +1,9 @@
+export function StationsPage() {
+  return (
+    <div>
+      <h1>
+        <b>Stations Page</b>
+      </h1>
+    </div>
+  );
+}
