--- src/renderer/src/routes/database.lazy.tsx
@@ -1,6 +1,0 @@
-import { createLazyFileRoute } from "@tanstack/react-router";
-import { DBsettingsHub } from "~/features/DBsettingsHub/DBsettingsHub";
-
-export const Route = createLazyFileRoute("/database")({
-  component: DBsettingsHub
-});
