+++ src/renderer/src/routes/stations.lazy.tsx
@@ -1,0 +1,6 @@
+import { createLazyFileRoute } from "@tanstack/react-router";
+import { StationsPage } from "~/features/StationsPage/StationsPage";
+
+export const Route = createLazyFileRoute("/stations")({
+  component: StationsPage
+});
